Add get help component to layout
Added component to the layout so that it is added
on all pages.

Copied styles and added to component SASS file.

Added JS to show the form when the button is
clicked. Form is hidden by default and when
loading the next page.

There is a gap where it does not show a success
message when the form is "submitted". Need to pick
up this with Robin Cox.