Fix single school DfE Sign in routing
If single school journeyType is set, then now link
to which school or trust are you buying for?
Options are dependant on the journeyType, single
schools see the trust and the school, MAT users
see everything.

In an ideal world MAT users would see some schools
within the trust but not all.